Add service category name to draft referrals
This is the first time we’ve thought about how the service category
information will be represented within a referral. This seems like a
fine starting point but the backend team might want to change it. We
need this information available so that we can populate the "What date
does the <service category name> service need to be completed by?" on
the completion deadline form.

Co-authored-by: George Eaton <george@dxw.com>