Add Cypress command to target a fieldset by legend
This allows us to select a radio button by text even when there are
multiple radios with the same text in different fieldsets in the same
page (common with e.g. “Yes” / “No”). This is especially important if
we’re going to take a testing-library[1]-esque approach of driving the
user interface by user-facing characteristics instead of by id
attributes, which I think we might.

[1] https://testing-library.com