Add an integration spec for creating and submitting a referral
This doesn’t use the real backend - everything is done with Wiremock. So
it’s not so much an end-to-end test, more of an integration test to
check that the UI is hooked up correctly.

I decided to use a testing-library [1] style approach of avoiding
targeting elements by IDs or other HTML attributes, instead preferring
to use user-facing attributes like text. I saw this recommendation
in some GitLab frontend testing advice [2], and, not having a strong opinion
myself, thought it worth a go.

I haven’t used any page objects here. I saw that we’ve used them here,
and saw them recommended in [3]. But, I tried writing some and thought
that it was kind of easier to tell what was going on as it is. Maybe
will change opinion on this as we write more tests.

[1] https://testing-library.com
[2] https://docs.gitlab.com/ee/development/testing_guide/frontend_testing.html#how-to-query-dom-elements
[3] https://sttp.site/chapters/testing-into-context/web-testing.html#page-objects