Add Desired Outcomes to activity presenter
We want to render them as part of the "add activities" form.

I initially expected we'd need to make an HTTP request to the
Interventions Service for this, but as they're stored on the service
category and we have their `id`s on the `SentReferral` object, we can do
this in memory.