Add textarea args to add activities view
This will render a govuk textarea for each desired outcome selected on
the referral.

To think about: should we pass the nunjucks macro into the view here,
like we did for the Intervention Details view? I initially went down
this route, but didn't want to pass things like the enclosing form input
into the view as well, so reverted to this implementation.

Still to come:
- Render any activities that have already been set above
the textarea.
- Validations on the texarea submit