WIP Add first test for community api service
Because of how we're instantiating RestClient rather than passing an
instance into the constructor, I've had to make the `restClient` method
private to be able to stub it in the test.

This probably suggests we should pass it in via dependency injection to
make it easier for testing, but I'd like to write all these tests before
we start refactoring so we have a safety net for any changes.