Add types and contracts for end of service report
This is based on the interventions service at commit 3263bed. The only
thing here that’s not in the interventions service is the `referralId`
property on the end of service report. I need this so that I can build
the end of service report journey using the same approach as we took for
action plan.