Add sessionDetailsSummary to CheckAnswersPresenter
This was originally taking the AttendancePresenter's
sessionDetailsSummary, but the CheckAnswersPresenter will now take the
FeedbackAnswersPresenter, so it needs its own method to display these
details.

I initially considered adding the sessionDetailsSummary to the
FeedbackAnswersPresenter, but as this is re-used across different user
journeys, the details displayed will change depending on where it's
used, so I've opted to give each Presenter its own version for now.